Software/requirements.txt
flask
paho-mqtt
psutil
pyserial
wmi
